server/requirements.txt
async-timeout==4.0.3
bidict==0.23.1
blinker==1.8.2
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
dnspython==2.6.1
eventlet==0.36.1
flask==3.0.3
Flask-Cors==4.0.1
Flask-MQTT==1.2.1
Flask-SocketIO==5.3.6
gevent==24.2.1
gevent-websocket==0.10.1
greenlet==3.0.3
h11==0.14.0
idna==3.7
importlib-metadata==8.0.0
itsdangerous==2.2.0
jinja2==3.1.4
MarkupSafe==2.1.5
paho-mqtt==1.6.1
python-dotenv==1.0.1
python-engineio==4.9.1
python-socketio==5.11.3
redis==5.0.7
requests==2.32.3
simple-websocket==1.0.0
typing-extensions==4.12.2
urllib3==2.2.2
werkzeug==3.0.3
wsproto==1.2.0
zipp==3.19.2
zope.event==5.0
zope.interface==6.4.post2
